fix: rename react-native-ci-cli to setup-ci after migration
