+++ .github/workflows/publish.yml
@@ -24,9 +24,3 @@ jobs:
         run: npm publish
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
-      - name: Build for alias 🏗️
-        run: yarn build:alias
-      - name: Publish package on NPM for alias 📦
-        run: npm publish
-        env:
-          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
