+++ CHANGELOG.md
@@ -1,46 +1,30 @@
-## [0.4.10](https://github.com/software-mansion-labs/react-native-ci-cli/compare/v0.4.9...v0.4.10) (2024-08-27)
-
-
-### Bug Fixes
-
-* add credits and spell check ([#84](https://github.com/software-mansion-labs/react-native-ci-cli/issues/84)) ([91a00e2](https://github.com/software-mansion-labs/react-native-ci-cli/commit/91a00e2a24207ce7535acdb2bbd9d542d9d41479))
-
-
-
-## [0.4.9](https://github.com/software-mansion-labs/react-native-ci-cli/compare/v0.4.8...v0.4.9) (2024-08-27)
-
+## [0.4.10](https://github.com/software-mansion/setup-ci/compare/v0.4.9...v0.4.10) (2024-08-27)
 
 ### Bug Fixes
 
-* add contents read permissions to detox workflows ([#83](https://github.com/software-mansion-labs/react-native-ci-cli/issues/83)) ([f6ab340](https://github.com/software-mansion-labs/react-native-ci-cli/commit/f6ab340ef66cda33727044e700c05a19f6ba2af9))
+- add credits and spell check ([#84](https://github.com/software-mansion/setup-ci/issues/84)) ([91a00e2](https://github.com/software-mansion-labs/react-native-ci-cli/commit/91a00e2a24207ce7535acdb2bbd9d542d9d41479))
 
-
-
-## [0.4.8](https://github.com/software-mansion-labs/react-native-ci-cli/compare/v0.4.7...v0.4.8) (2024-08-27)
-
+## [0.4.9](https://github.com/software-mansion/setup-ci/compare/v0.4.8...v0.4.9) (2024-08-27)
 
 ### Bug Fixes
 
-* add check read permissions to detox workflows ([#82](https://github.com/software-mansion-labs/react-native-ci-cli/issues/82)) ([6aa5079](https://github.com/software-mansion-labs/react-native-ci-cli/commit/6aa5079824fd0b5612b59150346ac0c6f564d564))
+- add contents read permissions to detox workflows ([#83](https://github.com/software-mansion/setup-ci/issues/83)) ([f6ab340](https://github.com/software-mansion-labs/react-native-ci-cli/commit/f6ab340ef66cda33727044e700c05a19f6ba2af9))
 
-
-
-## [0.4.7](https://github.com/software-mansion-labs/react-native-ci-cli/compare/v0.4.6...v0.4.7) (2024-08-26)
-
+## [0.4.8](https://github.com/software-mansion/setup-ci/compare/v0.4.7...v0.4.8) (2024-08-27)
 
 ### Bug Fixes
 
-* add repository url ([621c19a](https://github.com/software-mansion-labs/react-native-ci-cli/commit/621c19af3ac0fc63785b79f22c012ec4ae6bbb9b))
+- add check read permissions to detox workflows ([#82](https://github.com/software-mansion/setup-ci/issues/82)) ([6aa5079](https://github.com/software-mansion-labs/react-native-ci-cli/commit/6aa5079824fd0b5612b59150346ac0c6f564d564))
 
-
-
-## [0.4.6](https://github.com/software-mansion-labs/react-native-ci-cli/compare/v0.4.5...v0.4.6) (2024-08-26)
-
+## [0.4.7](https://github.com/software-mansion/setup-ci/compare/v0.4.6...v0.4.7) (2024-08-26)
 
 ### Bug Fixes
 
-* add token back ([a789a31](https://github.com/software-mansion-labs/react-native-ci-cli/commit/a789a315f947191572b47be70531b7e17329f45e))
-* remove unused token ([ca2d93c](https://github.com/software-mansion-labs/react-native-ci-cli/commit/ca2d93c138714c3e8bba9009ded175b6af56de1d))
+- add repository url ([621c19a](https://github.com/software-mansion/setup-ci/commit/621c19af3ac0fc63785b79f22c012ec4ae6bbb9b))
 
+## [0.4.6](https://github.com/software-mansion/setup-ci/compare/v0.4.5...v0.4.6) (2024-08-26)
 
+### Bug Fixes
 
+- add token back ([a789a31](https://github.com/software-mansion/setup-ci/commit/a789a315f947191572b47be70531b7e17329f45e))
+- remove unused token ([ca2d93c](https://github.com/software-mansion/setup-ci/commit/ca2d93c138714c3e8bba9009ded175b6af56de1d))
