+++ __tests__/cli-integration.test.ts
@@ -3,9 +3,7 @@ import { system, filesystem } from 'gluegun'
 const src = filesystem.path(__dirname, '..')
 
 const cli = async (cmd) =>
-  system.run(
-    'node ' + filesystem.path(src, 'bin', 'react-native-ci-cli') + ` ${cmd}`
-  )
+  system.run('node ' + filesystem.path(src, 'bin', 'setup-ci') + ` ${cmd}`)
 
 test('outputs version', async () => {
   const output = await cli('--version')
