--- docs/commands.md
@@ -1,3 +1,0 @@
-# Command Reference for react-native-ci-cli
-
-TODO: Add your command reference here
