+++ readme.md
@@ -2,9 +2,9 @@
 
 ## 🤔 Why?
 
-The problem we noticed is that setting up CI is performed once, needs to be researched every time, and is often trial and error driven. The process is repetitive. 
+The problem we noticed is that setting up CI is performed once, needs to be researched every time, and is often trial and error driven. The process is repetitive.
 
-We created a tool that bootstraps CI with the most used features, which are ready for customization in the future. When you run `npx setup-ci` React Native CI CLI generates GitHub workflows for the most popular CI tasks. 
+We created a tool that bootstraps CI with the most used features, which are ready for customization in the future. When you run `npx setup-ci` React Native CI CLI generates GitHub workflows for the most popular CI tasks.
 
 ## 📖 Usage
 
@@ -21,7 +21,7 @@ If your project has a monorepo structure, run the script from the root directory
 ## 💡 Demo
 
 Say we want to set up GitHub actions that run ESLint check and Jest tests on our project every time we create or push to an existing Pull Request.
-Let's try to use `react-native-ci-cli` to do the heavy lifting and generate all the necessary configurations.
+Let's try to use `setup-ci` to do the heavy lifting and generate all the necessary configurations.
 
 <p align="center">
   <img 
@@ -38,7 +38,7 @@ Alternatively, we can go fully automatic by passing flags `--preset --lint --jes
 - Monorepo structure is supported, but make sure you run the script from the app directory,
   not the monorepo root.
 
-When using `react-native-ci-cli`, you can provide additional flags to modify its default behavior.
+When using `setup-ci`, you can provide additional flags to modify its default behavior.
 
 <table>
   <tr>
@@ -90,7 +90,7 @@ The following are **feature flags** that can be used with `--preset` flag (they 
 
 ## 🔐 Repository secrets
 
-Some workflows generated by `react-native-ci-cli` require setting up repository secrets that can be then accessed by workflows as `${{ secrets.* }}`. Make sure to read logs printed by the script
+Some workflows generated by `setup-ci` require setting up repository secrets that can be then accessed by workflows as `${{ secrets.* }}`. Make sure to read logs printed by the script
 as you will always be prompted to create secrets if necessary. The following table shows a summary of secrets you might have to create for generated workflows to work properly.
 
 <table>
@@ -106,7 +106,7 @@ as you will always be prompted to create secrets if necessary. The following tab
 
 ## 💬 Your feedback
 
-This tool is supposed to be helpful to as many developers as possible - and therefore we are open to your ideas and general feedback! If you want to share your opinion about `react-native-ci-cli` or
+This tool is supposed to be helpful to as many developers as possible - and therefore we are open to your ideas and general feedback! If you want to share your opinion about `setup-ci` or
 have some thoughts about how it could be further developed, don't hesitate to create an issue or contact
 the maintainers directly.
 
@@ -127,7 +127,6 @@ React Native CI CLI is licensed under [The MIT License](LICENSE).
 This project has been built and is maintained thanks to the support from [Software Mansion](https://swmansion.com)
 
 [![swm](https://logo.swmansion.com/logo?color=white&variant=desktop&width=150&tag=react-native-reanimated-github 'Software Mansion')](https://swmansion.com)
-
 
 ## React Native CI CLI is created by Software Mansion
