+++ src/cli.ts
@@ -3,10 +3,10 @@ import { Options } from 'gluegun/build/types/domain/options'
 
 async function run(argv: string | Options) {
   const cli = build()
-    .brand('react-native-ci-cli')
+    .brand('setup-ci')
     .src(__dirname)
     .plugins('./node_modules', {
-      matching: 'react-native-ci-cli-*',
+      matching: 'setup-ci-*',
       hidden: true,
     })
     .defaultCommand(require('./commands/index'))
